Adding Authentication to Journal endpoints

+++ src/main/java/com/codeBlock/journalApp/service/JournalEntryService.java
@@ -31,7 +31,7 @@ public class JournalEntryService {
             JournalEntry saved = journalEntryRepository.save(journalEntry);
             user.getJournalEntries().add(saved);
             //user.setUserName(null); Added for checking transaction
-            userService.saveEntry(user);
+            userService.saveUser(user);
         }catch (Exception exception)
         {
             log.error("Exception : ",exception);
@@ -59,12 +59,23 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName)
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName)
     {
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed) {
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }catch (Exception exception)
+        {
+            System.out.println(exception);
+            throw new RuntimeException("An error occurred while deleting the entry. ",exception);
+        }
+        return removed;
     }
 
 }
